Added logging and changed labeler output to a dataframe.

+++ trane/ops/transformation_ops.py
@@ -1,5 +1,6 @@
 from ..utils.table_meta import TableMeta as TM
 from .op_base import OpBase
+import numpy
 
 TRANSFORMATION_OPS = ["IdentityTransformationOp", "DiffTransformationOp", "ObjectFrequencyTransformationOp"]
 __all__ = ["TransformationOpBase", "TRANSFORMATION_OPS"] + TRANSFORMATION_OPS
@@ -20,7 +21,7 @@ class DiffTransformationOp(TransformationOpBase):
     def execute(self, dataframe):
         index = dataframe.index
         for i in range(len(index) - 1, 0, -1):
-            dataframe.at[index[i], self.column_name] -= dataframe.at[index[i-1], self.column_name]
+            dataframe.at[index[i], self.column_name] -= dataframe.at[index[i-1], self.column_name].astype(numpy.float32)
         dataframe.at[index[0], self.column_name] = 0
         return dataframe
